--- src/main/java/ru/practicum/shareit/booking/dao/BookingDao.java
@@ -1,7 +1,0 @@
-package ru.practicum.shareit.booking.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.shareit.booking.model.Booking;
-
-public interface BookingDao extends JpaRepository<Booking, Long> {
-}
