+++ src/main/java/ru/practicum/shareit/item/repository/ItemRepository.java
@@ -1,11 +1,11 @@
-package ru.practicum.shareit.item.dao;
+package ru.practicum.shareit.item.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
 
-public interface ItemDao extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
     List<Item> findAllByOwnerId(long ownerId);
     List<Item> findAllByNameOrDescriptionContainingIgnoreCase(String nameSearch, String descriptionSearch);
 }
