+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -1,11 +1,10 @@
-package ru.practicum.shareit.user.dao;
+package ru.practicum.shareit.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.shareit.user.model.User;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
