+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -3,7 +3,7 @@ package ru.practicum.shareit.user.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.user.dao.UserDao;
+import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.util.exception.AlreadyExistException;
 import ru.practicum.shareit.util.exception.NotFoundException;
@@ -14,13 +14,13 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
-    private final UserDao userDao;
+    private final UserRepository repository;
 
     @Override
     public User create(User user) {
         throwIfEmailExist(user.getEmail());
 
-        User userCreated = userDao.save(user);
+        User userCreated = repository.save(user);
 
         log.info("Добавлен новый пользователь: {}", userCreated);
         return userCreated;
@@ -28,7 +28,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User getById(long id) {
-        User user = userDao.findById(id)
+        User user = repository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id " + id + " не найден"));
 
         log.info("Передан пользователь: {}", user);
@@ -37,7 +37,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User update(long id, User user) {
-        User userToUpdate = userDao.findById(id)
+        User userToUpdate = repository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id " + id + " не найден"));
 
         if (user.getName() != null) {
@@ -50,7 +50,7 @@ public class UserServiceImpl implements UserService {
             userToUpdate.setEmail(user.getEmail());
         }
 
-        User userUpdated = userDao.save(userToUpdate);
+        User userUpdated = repository.save(userToUpdate);
 
         log.info("Обновлён пользователь: {}", userUpdated);
         return userUpdated;
@@ -58,23 +58,23 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(long id) {
-        User user = userDao.findById(id)
+        User user = repository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Пользователь с id " + id + " не найден"));
 
-        userDao.deleteById(id);
+        repository.deleteById(id);
         log.info("Удалён пользователь: {}", user);
     }
 
     @Override
     public List<User> getAll() {
-        List<User> users = userDao.findAll();
+        List<User> users = repository.findAll();
 
         log.info("Передан список всех пользователей");
         return users;
     }
 
     private void throwIfEmailExist(String email) {
-        if (userDao.findByEmail(email).isPresent()) {
+        if (repository.findByEmail(email).isPresent()) {
             throw new AlreadyExistException("Email " + email + " уже существует");
         }
     }
