+++ src/main/resources/application.yml
@@ -11,12 +11,16 @@ spring:
     password: test
     username: test
     url: jdbc:h2:mem:shareit
-  config:
-    activate:
-      on-profile: ci,test
+#  config:
+#    activate:
+#      on-profile: ci,test
   sql:
     init:
       mode: always
+  h2:
+    console:
+      enabled: true
+
 logging:
   level:
     org:
